pyproject.toml
[tool.poetry]
name = "akari-bot"
version = "0.1.0"
description = ""
authors = [
    "OasisAkari <oasisakari@gmail.com>",
    "Dianliang233 <dianliang@teahou.se>",
    "DoroWolf <dorowolf314@gmail.com>",
    "nullcat <null@teahou.se>",
    "_LittleC_ <i@ltlec.com>",
    "cloudw233 <cloudw233@outlook.com>",
]
license = "AGPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.0"

# akari-bot-service
akari-bot-webrender = "0.5.12"
akari-bot-webui = "0.3.6"

# lib
coverage = "^7.9.1"
cryptography = "^46.0.0"
deprecated = "^1.2.18"
poetry-plugin-export = "^1.9.0"
pre-commit = "^4.0.0"
pycryptodome = "^3.18.0"
python-multipart = "^0.0.21"
pywin32 = { version = "^308", platform = "win32" }

# bots
aiocqhttp = "^1.4.3"
aiogram = "^3.4.1"
argon2-cffi = "^25.0.0"
fastapi = ">=0.120.0, <1.0.0"
ffmpy = "^1.0.0"
khl-py = "^0.3.16"
matrix-nio = "^0.25.0"
py-cord = {git = "https://github.com/Pycord-Development/pycord"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
qq-botpy = "^1.2.1"
slowapi = "^0.1.9"
uvicorn = {extras = ["standard"], version = ">=0.38.0, <1.0.0"}

# core
cattrs = "^25.0.0"
filetype = "^1.2.0"
japanera = "^2.1.3"
orjson = "^3.10.9"
Pillow = "^12.0.0"
tenacity = "^9.0.0"

# core.config
tomlkit = "^0.14.0"

# core.database
tortoise-orm = {extras = ["asyncmy"], version = "^0.25.0"}

# core.http
aiofile = "^3.8.1"
attrs = "^25.0.0"
httpx = "^0.28.1"
jinja2 = "^3.1.6"

# core.logger
loguru = "^0.7.0"

# core.utils
ff3 = "^1.0.2"
psutil = "^7.0.0"
tabulate = "^0.9.0"

# core.scheduler
APScheduler = "^3.9.1-post.0"

# modules.core
matplotlib = "^3.6.3"
py-cpuinfo = "^9.0.0"

# modules
beautifulsoup4 = "^4.11.1"
emoji = "^2.12.1"
feedparser = "^6.0.10"
google-play-scraper = "^1.2.2"
gql = "^4.0.0"
langconv = {git = "https://github.com/Teahouse-Studios/langconv.py.git"}
mcstatus = "^12.0.0"
numpy = "<2.4.2"
openai = "^2.0.0"
pandas = "^2.2.0"
pint = "^0.25.0"
python-whois = "^0.9.0"
rdkit = "^2025.0.0"
simpleeval = "^1.0.0"
webcolors = "^25.0.0"
wolframalpha = "^5.0.0"

# uncategorized

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pycodestyle = "2.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3

[tool.pyright]
reportUndefinedVariable = "error"
